+++ typespec-tests/src/main/java/com/_specs_/azure/core/basic/BasicClient.java
@@ -483,7 +483,7 @@ public final class BasicClient {
         // Generated convenience method for createOrUpdateWithResponse
         RequestOptions requestOptions = new RequestOptions();
         JsonMergePatchHelper.getUserAccessor().prepareModelForJsonMergePatch(resource, true);
-        BinaryData resourceInBinaryData = BinaryData.fromString(BinaryData.fromObject(resource).toString());
+        BinaryData resourceInBinaryData = BinaryData.fromBytes(BinaryData.fromObject(resource).toBytes());
         JsonMergePatchHelper.getUserAccessor().prepareModelForJsonMergePatch(resource, false);
         return createOrUpdateWithResponse(id, resourceInBinaryData, requestOptions).getValue().toObject(User.class);
     }