+++ typespec-tests/src/test/java/com/type/property/additionalproperties/ExtendsTests.java
@@ -9,6 +9,8 @@ import com.type.property.additionalproperties.models.ExtendsModelAdditionalPrope
 import com.type.property.additionalproperties.models.ExtendsModelArrayAdditionalProperties;
 import com.type.property.additionalproperties.models.ExtendsStringAdditionalProperties;
 import com.type.property.additionalproperties.models.ExtendsUnknownAdditionalProperties;
+import com.type.property.additionalproperties.models.ExtendsUnknownAdditionalPropertiesDerived;
+import com.type.property.additionalproperties.models.ExtendsUnknownAdditionalPropertiesDiscriminatedDerived;
 import com.type.property.additionalproperties.models.ModelForRecord;
 
 import java.util.Arrays;
@@ -22,6 +24,8 @@ public class ExtendsTests {
     private final ExtendsModelClient extendsModelClient = new AdditionalPropertiesClientBuilder().buildExtendsModelClient();
     private final ExtendsStringClient extendsStringClient = new AdditionalPropertiesClientBuilder().buildExtendsStringClient();
     private final ExtendsUnknownClient extendsUnknownClient = new AdditionalPropertiesClientBuilder().buildExtendsUnknownClient();
+    private final ExtendsUnknownDerivedClient extendsUnknownDerivedClient = new AdditionalPropertiesClientBuilder().buildExtendsUnknownDerivedClient();
+    private final ExtendsUnknownDiscriminatedClient extendsUnknownDiscriminatedClient = new AdditionalPropertiesClientBuilder().buildExtendsUnknownDiscriminatedClient();
 
     @Test
     public void testExtendsFloat() {
@@ -103,5 +107,44 @@ public class ExtendsTests {
         Assertions.assertNotNull(properties.getAdditionalProperties());
         Assertions.assertEquals("ExtendsUnknownAdditionalProperties", properties.getName());
         Assertions.assertIterableEquals(propertyMap.entrySet(), properties.getAdditionalProperties().entrySet());
+    }
+
+    @Test
+    @Disabled("`AdditionalProperties` is not json serialized, so body provided doesn't match expected body.")
+    public void testExtendsUnknownDerivedClient() {
+        Map<String, Object> additionalProperty = new LinkedHashMap<>();
+        additionalProperty.put("prop1", 32);
+        additionalProperty.put("prop2", true) ;
+        additionalProperty.put("prop3", "abc") ;
+        ExtendsUnknownAdditionalPropertiesDerived body =
+                new ExtendsUnknownAdditionalPropertiesDerived("ExtendsUnknownAdditionalProperties", 314);
+        body.setAge(2.71828);
+        body.setAdditionalProperties(additionalProperty);
+        extendsUnknownDerivedClient.put(body);
+
+        ExtendsUnknownAdditionalPropertiesDerived properties = extendsUnknownDerivedClient.get();
+        Assertions.assertNotNull(properties);
+        Assertions.assertNotNull(properties.getAdditionalProperties());
+        Assertions.assertEquals("ExtendsUnknownAdditionalProperties", properties.getName());
+        Assertions.assertIterableEquals(additionalProperty.entrySet(), properties.getAdditionalProperties().entrySet());
+    }
+
+    @Test
+    @Disabled("`AdditionalProperties` is not json serialized, so body provided doesn't match expected body.")
+    public void testExtendsUnknownDiscriminatedClient() {
+        Map<String, Object> additionalProperty = new LinkedHashMap<>();
+        additionalProperty.put("prop1", 32);
+        additionalProperty.put("prop2", true) ;
+        additionalProperty.put("prop3", "abc") ;
+        ExtendsUnknownAdditionalPropertiesDiscriminatedDerived body = new ExtendsUnknownAdditionalPropertiesDiscriminatedDerived("Derived", 314);
+        body.setAge(2.71828);
+        body.setAdditionalProperties(additionalProperty);
+        extendsUnknownDiscriminatedClient.put(body);
+
+        ExtendsUnknownAdditionalPropertiesDiscriminatedDerived properties = (ExtendsUnknownAdditionalPropertiesDiscriminatedDerived) extendsUnknownDiscriminatedClient.get();
+        Assertions.assertNotNull(properties);
+        Assertions.assertNotNull(properties.getAdditionalProperties());
+        Assertions.assertEquals("Derived", properties.getName());
+        Assertions.assertIterableEquals(additionalProperty.entrySet(), properties.getAdditionalProperties().entrySet());
     }
 }